[fix] fix the type missmatch (usize -> u32)

+++ src/cpu.rs
@@ -76,9 +76,9 @@ impl CPU {
         println!("new epc:0x{:x}", self.pc);
     }
 
-    pub fn trans_addr(&mut self, addr: i32) -> Option<usize> {
+    pub fn trans_addr(&mut self, addr: i32) -> Option<u32> {
         let base_addr = self.bus.dram.base_addr;
-        match self.mmu.trans_addr(addr as usize, 
+        match self.mmu.trans_addr(addr as u32, 
                                   self.csrs.read(CSRname::satp.wrap()), 
                                   &self.bus.dram, &self.priv_lv) {
             Ok(addr) => {
